backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
httpx==0.28.1
python-multipart==0.0.20
boto3==1.38.12
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.4.0
pydantic[email]==2.10.6
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1